+++ com.unity.render-pipelines.high-definition/HDRP/RenderPipeline/HDStringConstants.cs
@@ -247,6 +247,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public static readonly int _InvViewMatrixStereo = Shader.PropertyToID("_InvViewMatrixStereo");
         public static readonly int _InvProjMatrixStereo = Shader.PropertyToID("_InvProjMatrixStereo");
         public static readonly int _InvViewProjMatrixStereo = Shader.PropertyToID("_InvViewProjMatrixStereo");
+        public static readonly int _PrevViewProjMatrixStereo = Shader.PropertyToID("_PrevViewProjMatrixStereo");
+        public static readonly int _TextureWidthScaling = Shader.PropertyToID("_TextureWidthScaling"); // 0.5 for SinglePassDoubleWide (stereo) and 1.0 otherwise
 
         public static readonly int _DepthTexture                   = Shader.PropertyToID("_DepthTexture");
         public static readonly int _CameraColorTexture             = Shader.PropertyToID("_CameraColorTexture");