+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed an issue with SAMPLE_TEXTURECUBE_SHADOW macro
 - Fixed issue with color picker not display correctly when game and scene view are visible at the same time
 - Fixed an issue with reflection probe face fading
+- Fixed camera motion vectors shader and associated matrices to update correctly for single-pass double-wide stereo rendering
 
 ### Changed
 - Changed the way depth & color pyramids are built to be faster and better quality, thus improving the look of distortion and refraction.